pyproject.toml
[project]
name = "cloud-run-functions"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "onikarubi", email = "printuserselect@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.3.0",
    "click>=8.2.1",
    "fastapi>=0.115.12",
    "google-api-python-client>=2.170.0",
    "google-auth>=2.40.2",
    "google-auth-oauthlib>=1.2.2",
    "httpx>=0.28.1",
    "pydantic-settings>=2.9.1",
    "python-jose>=3.4.0",
    "uvicorn>=0.34.2",
]

[project.scripts]
cloud-run-functions = "app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.11",
]
[tool.pytest.ini_options]
addopts = "-p no:warnings --cov --cov-report=term-missing"
log_cli = true
pythonpath = "./src"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 135
src = ["src", "tests"]
exclude = [".venv"]

[tool.ruff.lint]
select = [
    "I",
    "F",
    "E",
    "S",
    "N",
    "B"
]
ignore = [
    "F401",
    "E501",
    "S101"
]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true

[tool.ruff.flake8-bugbear]
extend-immutable-calls = ["fastapi.Depends", "fastapi.params.Depends", "fastapi.Query", "fastapi.params.Query"]
